feat: Implement dual strategy calculation and refactor for quality
- Implemented distinct parameter calculation for Directional vs Simultaneous strategies in 'bidi_calculator.rs'.
- Updated 'RetroAnalysisResults.vue' and 'StraddleSimultaneousCard.vue' to display specific simultaneous parameters.
- Fixed archive synchronization issue by including new parameters in archive data.
- Refactored 'calendar_import_commands.rs' to use 'calendar_db_helper.rs' to reduce file size.
- Refactored 'MetaAnalysisModal.vue' by extracting sub-components ('DivergenceGraph', 'ProfitabilityMatrix', 'LeaderboardTable', 'OptimizerTable') to reduce file size.
- Fixed 'unwrap()' usage in 'projection.rs'.
- Fixed frontend quality issues (console.error, alert, as any) in 'RetroactiveAnalysisView.vue', 'PlanningView.vue', and PDF generators.
- Passed full code audit and regression checks.

+++ src-tauri/src/services/economic_event_loader.rs
@@ -126,9 +126,9 @@ impl EconomicEventLoader {
             })?;
 
         // Parse optional numeric values
-        let actual = record.get(5).and_then(|s| s.parse::<f64>().ok());
-        let forecast = record.get(6).and_then(|s| s.parse::<f64>().ok());
-        let previous = record.get(7).and_then(|s| s.parse::<f64>().ok());
+        let actual = record.get(5).and_then(|s| s.parse::<f32>().ok());
+        let forecast = record.get(6).and_then(|s| s.parse::<f32>().ok());
+        let previous = record.get(7).and_then(|s| s.parse::<f32>().ok());
 
         Ok(NewCalendarEvent {
             symbol: symbol.to_string(),
